requirements.txt
streamlit
openai
chromadb
pandas
openpyxl
semantic-router
sentence-transformers
python-dotenv
groq
sqlalchemy
torch
transformers
huggingface-hub